chore: create ryot_app module for plugins and bundles
