+++ crates/ryot/src/bevy_ryot/game/mod.rs
@@ -1,12 +1,1 @@
-use bevy::prelude::*;
-use ryot_tiled::prelude::*;
-
 pub mod tile_flags;
-
-pub struct GamePlugin;
-
-impl Plugin for GamePlugin {
-    fn build(&self, app: &mut App) {
-        app.add_event::<LoadObjects>().add_plugins(ElevationPlugin);
-    }
-}
