+++ crates/ryot_assets/src/lib.rs
@@ -5,7 +5,6 @@ use bevy_ecs::prelude::Resource;
 
 pub mod atlas;
 pub mod catalog;
-pub mod plugins;
 pub mod sprites;
 pub mod visual_elements;
 
@@ -15,8 +14,6 @@ pub mod prelude {
     pub use crate::{
         atlas::AtlasLayoutsAsset,
         catalog::{prepare_sprite_sheets, Catalog, CatalogAsset},
-        content_plugin,
-        plugins::{BaseContentPlugin, MetaContentPlugin, VisualContentPlugin},
         sprites::{prepare_sprite_layouts, prepare_sprite_meshes},
         visual_elements::{prepare_visual_elements, VisualElementsAsset},
     };
