+++ crates/ryot_internal/src/prelude.rs
@@ -1,14 +1,19 @@
-pub use ryot_assets::prelude::*;
 pub use ryot_content::prelude::*;
 pub use ryot_core::prelude::*;
 pub use ryot_sprites::prelude::*;
 pub use ryot_tiled::prelude::*;
 
-#[cfg(feature = "ryot_tibia")]
-pub use ryot_tibia::prelude as tibia;
+#[cfg(feature = "ryot_app")]
+pub use ryot_app::prelude::*;
+
+#[cfg(feature = "ryot_assets")]
+pub use ryot_assets::prelude::*;
 
 #[cfg(feature = "ryot_pathfinder")]
 pub use ryot_pathfinder::prelude::*;
 
 #[cfg(feature = "ryot_ray_casting")]
 pub use ryot_ray_casting::prelude::*;
+
+#[cfg(feature = "ryot_tibia")]
+pub use ryot_tibia::prelude as tibia;
